==> 1.Exploratory_Data_Analysis/notebooks/8.0_create_feature_index.ipynb <==
import argparse
import pathlib

import pyarrow.parquet as pq

# parse command line arguments
parser = argparse.ArgumentParser()
parser.add_argument("--cell_type", help="cell type to analyze", type=str, default="all")

args = parser.parse_args()
cell_type = args.cell_type

# Import Data
# set data file path under pathlib path for multi-system use
file_path = pathlib.Path(f"../../data/{cell_type}_preprocessed_sc_norm.parquet")
columns_list = pq.read_schema(file_path).names
columns_list = [x for x in columns_list if "Metadata" not in x]
columns_list = [x for x in columns_list if "__index_level_0__" not in x]

# index output for features
output_file = pathlib.Path(f"../features/{cell_type}_feature_index.txt")
# create output directory if it doesn't exist
output_file.parent.mkdir(parents=True, exist_ok=True)

# write each feature to a file
with open(output_file, "w") as f:
    for item in columns_list:
        f.write("%s\n" % item)
